Move maps interfaces to geo module
These could eventually move to their own `maps` module but moving
the map fragment implementations would be very cumbersome and is
probably worth doing a single dedicated piece of work.
