+++ collect_app/src/main/java/org/odk/collect/android/geo/GoogleMapFragment.java
@@ -54,7 +54,8 @@ import org.odk.collect.android.utilities.IconUtils;
 import org.odk.collect.android.utilities.MapFragmentReferenceLayerUtils;
 import org.odk.collect.android.utilities.ThemeUtils;
 import org.odk.collect.androidshared.utils.ToastUtils;
-import org.odk.collect.geo.MapPoint;
+import org.odk.collect.geo.maps.MapFragment;
+import org.odk.collect.geo.maps.MapPoint;
 import org.odk.collect.location.GoogleFusedLocationClient;
 import org.odk.collect.location.LocationClient;
 import org.odk.collect.location.LocationClientProvider;
@@ -71,7 +72,7 @@ import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import timber.log.Timber;
 
 public class GoogleMapFragment extends SupportMapFragment implements
-    MapFragment, LocationListener, LocationClient.LocationClientListener,
+        MapFragment, LocationListener, LocationClient.LocationClientListener,
     GoogleMap.OnMapClickListener, GoogleMap.OnMapLongClickListener,
     GoogleMap.OnMarkerClickListener, GoogleMap.OnMarkerDragListener,
     GoogleMap.OnPolylineClickListener {
