+++ collect_app/src/main/java/org/odk/collect/android/geo/MapboxMapConfigurator.java
@@ -13,6 +13,7 @@ import com.google.common.collect.ImmutableSet;
 import org.odk.collect.android.R;
 import org.odk.collect.android.preferences.PrefUtils;
 import org.odk.collect.androidshared.utils.ToastUtils;
+import org.odk.collect.geo.maps.MapFragment;
 import org.odk.collect.shared.Settings;
 
 import java.io.File;
