+++ collect_app/src/main/java/org/odk/collect/android/geo/OsmDroidMapConfigurator.java
@@ -9,6 +9,7 @@ import com.google.common.collect.ImmutableSet;
 
 import org.odk.collect.android.R;
 import org.odk.collect.android.preferences.PrefUtils;
+import org.odk.collect.geo.maps.MapFragment;
 import org.odk.collect.shared.Settings;
 
 import java.io.File;
