+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -67,7 +67,7 @@ import org.odk.collect.android.gdrive.GoogleAccountPicker;
 import org.odk.collect.android.gdrive.GoogleAccountsManager;
 import org.odk.collect.android.gdrive.GoogleApiProvider;
 import org.odk.collect.android.geo.DirectoryReferenceLayerRepository;
-import org.odk.collect.android.geo.MapFragmentFactory;
+import org.odk.collect.geo.maps.MapFragmentFactory;
 import org.odk.collect.android.geo.MapProvider;
 import org.odk.collect.android.geo.ReferenceLayerRepository;
 import org.odk.collect.android.instancemanagement.InstanceAutoSender;
