+++ collect_app/src/main/java/org/odk/collect/android/utilities/MapFragmentReferenceLayerUtils.kt
@@ -1,8 +1,8 @@
 package org.odk.collect.android.utilities
 
 import android.os.Bundle
-import org.odk.collect.android.geo.MapFragment
 import org.odk.collect.android.geo.ReferenceLayerRepository
+import org.odk.collect.geo.maps.MapFragment
 import java.io.File
 
 object MapFragmentReferenceLayerUtils {
