+++ collect_app/src/test/java/org/odk/collect/android/geo/TestMapFragment.java
@@ -6,7 +6,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.FragmentActivity;
 
-import org.odk.collect.geo.MapPoint;
+import org.odk.collect.geo.maps.MapFragment;
+import org.odk.collect.geo.maps.MapPoint;
 
 import java.util.HashMap;
 import java.util.List;
