+++ geo/src/main/java/org/odk/collect/geo/GeoUtils.java
@@ -2,6 +2,7 @@ package org.odk.collect.geo;
 
 import android.location.Location;
 
+import org.odk.collect.geo.maps.MapPoint;
 import org.odk.collect.shared.strings.StringUtils;
 
 import java.util.List;
