+++ geo/src/main/java/org/odk/collect/geo/maps/MapFragmentFactory.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.geo
+package org.odk.collect.geo.maps
 
 import android.content.Context
