+++ geo/src/main/java/org/odk/collect/geo/maps/MapPoint.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package org.odk.collect.geo;
+package org.odk.collect.geo.maps;
 
 import android.os.Parcel;
 import android.os.Parcelable;
